# Couple of small perf tweaks