dotnet C:\GitHub\azure-sdk-for-net\artifacts\bin\Azure.Storage.Blobs.Perf\Release\net6.0\Azure.Storage.Blobs.Perf.dll DownloadBlob --size 16384 --parallel 1 --warmup 60 --duration 60

Before:
Completed 1,756 operations in a weighted-average of 59.97s (29.28 ops/s, 0.03415 s/op, 0.23% CPU)

After:
Completed 1,836 operations in a weighted-average of 59.99s (30.60 ops/s, 0.03268 s/op, 0.059% CPU)